// src/entities/openapi/model/dataProductDtoStateCode.ts
/**
 * agridata.ch API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * State of the data product
 */
export const DataProductDtoStateCode = {
    Draft: 'DRAFT',
    Active: 'ACTIVE'
} as const;
export type DataProductDtoStateCode = typeof DataProductDtoStateCode[keyof typeof DataProductDtoStateCode];